Separate public interfaces
- Redefine some interfaces as public
